Use composition instead of inheritance for Bundle; Move Bundle to separate file; Make Resources struct again
