+++ Sources/RswiftCore/ResourceTypes/Resources.swift
@@ -14,7 +14,7 @@ enum ResourceParsingError: Error {
   case parsingFailed(String)
 }
 
-class Resources {
+struct Resources {
   let assetFolders: [AssetFolder]
   let images: [Image]
   let fonts: [Font]
@@ -56,7 +56,7 @@ class Resources {
       if let resourceFile = tryResourceParsing({ try ResourceFile(url: url) }) {
         if resourceFile.pathExtension == "bundle" {
           let bundle = Bundle(bundleUrl: url, fileManager: fileManager)
-          if bundle.resourceFiles.count != 0 {
+          if bundle.resources.resourceFiles.count != 0 {
             bundles.append(bundle)
           }
         } else {
@@ -79,70 +79,6 @@ class Resources {
   }
 }
 
-class Bundle: Resources {
-  let bundleName: String
-
-  init(bundleUrl: URL, fileManager: FileManager) {
-    let resourceURLs = Bundle.resourceURLs(for: bundleUrl, fileManager: fileManager)
-
-    let bundleNameWithExtension = bundleUrl.lastPathComponent
-    bundleName = bundleNameWithExtension.replacingOccurrences(of: ".bundle", with: "")
-
-    super.init(resourceURLs: resourceURLs, fileManager: fileManager)
-  }
-  
-  fileprivate static func resourceURLs(for bundleUrl: URL, fileManager: FileManager) -> [URL] {
-    var resourceURLs = [URL]()
-    
-    let resourceDirectoryTypes: [WhiteListedExtensionsResourceType.Type] = [AssetFolder.self]
-    var resourceDirectorySuffixes = Set<String>()
-    resourceDirectoryTypes.forEach { resourceType in
-      resourceType.supportedExtensions.forEach { ext in
-        resourceDirectorySuffixes.insert(ext)
-      }
-    }
-
-    var prefixesToSkip = Set<String>()
-    prefixesToSkip.insert(".") // Ignore files like .DS_Store, etc.
-    
-    let isResourceDirectory: ((URL) -> Bool) = { url in
-      for suffix in resourceDirectorySuffixes {
-        if url.absoluteString.hasSuffix(suffix + "/") {
-          return true
-        }
-      }
-      return false
-    }
-    
-    let shouldSkip: ((URL) -> Bool) = { url in
-      let name = url.lastPathComponent
-      for prefix in prefixesToSkip {
-        if name.hasPrefix(prefix) {
-          return true
-        }
-      }
-      return false
-    }
-    
-    guard let enumerator = fileManager.enumerator(at: bundleUrl, includingPropertiesForKeys: nil) else {
-      return []
-    }
-
-    for case let itemURL as URL in enumerator {
-      if itemURL.isDirectory() {
-        if isResourceDirectory(itemURL) {
-          resourceURLs.append(itemURL)
-          enumerator.skipDescendents()
-        }
-      } else if !shouldSkip(itemURL) {
-        resourceURLs.append(itemURL)
-      }
-    }
-    
-    return resourceURLs
-  }
-}
-
 private func tryResourceParsing<T>(_ parse: () throws -> T) -> T? {
   do {
     return try parse()
@@ -153,17 +89,5 @@ private func tryResourceParsing<T>(_ parse: () throws -> T) -> T? {
     return nil
   } catch {
     return nil
-  }
-}
-
-fileprivate extension URL {
-  func isDirectory() -> Bool {
-    do {
-      let value = try resourceValues(forKeys:[.isDirectoryKey])
-      return value.isDirectory ?? false
-    }
-    catch {
-      return false
-    }
   }
 }
